Moving one integration style test from ServerConnectionTest to new file
