+++ geode-core/src/test/java/org/apache/geode/internal/cache/tier/sockets/ServerConnectionTest.java
@@ -161,115 +161,4 @@ public class ServerConnectionTest {
         .isExactlyInstanceOf(AuthenticationRequiredException.class)
         .hasMessage(HandShake_NO_SECURITY_CREDENTIALS_ARE_PROVIDED.getRawText());
   }
-
-  class TestMessage extends Message {
-    private final Lock lock = new ReentrantLock();
-    private final Condition testGate = lock.newCondition();
-    private boolean signalled = false;
-
-    public TestMessage() {
-      super(3, Version.CURRENT);
-      messageType = MessageType.REQUEST;
-      securePart = new Part();
-    }
-
-    @Override
-    public void receive() throws IOException {
-      try {
-        lock.lock();
-        testGate.await(10, TimeUnit.SECONDS);
-      } catch (InterruptedException e) {
-        throw new RuntimeException(e);
-      } finally {
-        lock.unlock();
-        if (!signalled) {
-          fail("Message never received continueProcessing call");
-        }
-      }
-    }
-
-    public void continueProcessing() {
-      lock.lock();
-      testGate.signal();
-      signalled = true;
-      lock.unlock();
-    }
-  }
-
-  class TestServerConnection extends OriginalServerConnection {
-
-    private TestMessage testMessage;
-
-    /**
-     * Creates a new <code>ServerConnection</code> that processes messages received from an edge
-     * client over a given <code>Socket</code>.
-     */
-    public TestServerConnection(Socket socket, InternalCache internalCache,
-        CachedRegionHelper helper, CacheServerStats stats, int hsTimeout, int socketBufferSize,
-        String communicationModeStr, byte communicationMode, Acceptor acceptor,
-        SecurityService securityService) {
-      super(socket, internalCache, helper, stats, hsTimeout, socketBufferSize, communicationModeStr,
-          communicationMode, acceptor, securityService);
-
-      setClientDisconnectCleanly(); // Not clear where this is supposed to be set in the timeout
-                                    // path
-    }
-
-    @Override
-    protected void doHandshake() {
-      ClientProxyMembershipID proxyID = mock(ClientProxyMembershipID.class);
-      when(proxyID.getDistributedMember()).thenReturn(mock(InternalDistributedMember.class));
-      ServerSideHandshake handshake = mock(ServerSideHandshake.class);
-      when(handshake.getMembershipId()).thenReturn(proxyID);
-      when(handshake.getVersion()).thenReturn(Version.CURRENT);
-
-      setHandshake(handshake);
-      setProxyId(proxyID);
-
-      processHandShake();
-      initializeCommands();
-
-      setFakeRequest();
-
-      long fakeId = -1;
-      MessageIdExtractor extractor = mock(MessageIdExtractor.class);
-      when(extractor.getUniqueIdFromMessage(getRequestMessage(), handshake.getEncryptor(),
-          Connection.DEFAULT_CONNECTION_ID)).thenReturn(fakeId);
-      setMessageIdExtractor(extractor);
-    }
-
-    @Override
-    void handleTermination(boolean timedOut) {
-      super.handleTermination(timedOut);
-      testMessage.continueProcessing();
-    }
-
-    private void setFakeRequest() {
-      testMessage = new TestMessage();
-      setRequestMsg(testMessage);
-    }
-  }
-
-  /**
-   * This test sets up a TestConnection which will register with the ClientHealthMonitor and then
-   * block waiting to receive a fake message. This message will arrive just after the health monitor
-   * times out this connection and kills it. The test then makes sure that the connection correctly
-   * handles the terminated state and exits.
-   */
-  @Test
-  public void terminatingConnectionHandlesNewRequestsGracefully() throws Exception {
-    when(cache.getCacheTransactionManager()).thenReturn(mock(TXManagerImpl.class));
-    ClientHealthMonitor.createInstance(cache, 100, mock(CacheClientNotifierStats.class));
-    ClientHealthMonitor clientHealthMonitor = ClientHealthMonitor.getInstance();
-    when(acceptor.getClientHealthMonitor()).thenReturn(clientHealthMonitor);
-    when(acceptor.getConnectionListener()).thenReturn(mock(ConnectionListener.class));
-    when(securityService.isIntegratedSecurity()).thenReturn(true);
-
-    TestServerConnection testServerConnection =
-        new TestServerConnection(socket, cache, mock(CachedRegionHelper.class), stats, 0, 0, null,
-            CommunicationMode.PrimaryServerToClient.getModeNumber(), acceptor, securityService);
-    MockitoAnnotations.initMocks(this);
-
-    testServerConnection.run();
-  }
 }
